Census data joiner with complete GUI
This is what Shiyan and I did, a complete GUI for join census data to Shp. This program can pick up different data folder and join them together before joining to the shp. User have to choose the right GEOID name for the shp, because the GEOID fieldname is also changing among different Shps.